refactor: move setup from utils
